refactor: relocated utility methods
