+++ src/integrations/Clevertap/browser.js
@@ -2,7 +2,8 @@
 import get from 'get-value';
 import logger from '../../utils/logUtil';
 import ScriptLoader from '../../utils/ScriptLoader';
-import { extractCustomFields, getDefinedTraits, isArray, isObject } from '../../utils/utils';
+import { extractCustomFields, isArray, isObject } from '../../utils/utils';
+import { getDefinedTraits } from '../../utils/commonUtils';
 import { NAME } from './constants';
 
 class Clevertap {
