+++ src/integrations/GA4/utils.js
@@ -1,7 +1,7 @@
 import { eventNamesConfigArray, itemParametersConfigArray } from './ECommerceEventConfig';
 
 import { pageEventParametersConfigArray } from './PageEventConfig';
-import { type } from '../../utils/utils';
+import { type } from '../../utils/commonUtils';
 import logger from '../../utils/logUtil';
 
 /**
