+++ src/integrations/Sentry/browser.js
@@ -6,8 +6,8 @@
 import get from 'get-value';
 import logger from '../../utils/logUtil';
 import { SentryScriptLoader, sentryInit } from './utils';
-import { removeUndefinedAndNullValues } from '../../utils/commonUtils';
-import { getDefinedTraits, isObject } from '../../utils/utils';
+import { removeUndefinedAndNullValues, getDefinedTraits } from '../../utils/commonUtils';
+import { isObject } from '../../utils/utils';
 import { NAME } from './constants';
 
 class Sentry {
