+++ src/integrations/YandexMetrica/browser.js
@@ -4,7 +4,7 @@ import { ecommEventPayload, sendEvent, ecommerceEventMapping } from './utils';
 import { removeUndefinedAndNullValues, getHashFromArrayWithDuplicate } from '../../utils/commonUtils';
 import { NAME } from './constants';
 import { LOAD_ORIGIN } from '../../utils/ScriptLoader';
-import { getDefinedTraits } from '../../utils/utils';
+import { getDefinedTraits } from '../../utils/commonUtils';
 
 class YandexMetrica {
   constructor(config, analytics) {
